PS-10045: fix compiler warnings in kmipcore/kmipclient and tighten maintainer warning mode
https://perconadev.atlassian.net/browse/PS-10045

- add/cleanup maintainer warning profiles in kmipcore and kmipclient CMake
  (target-scoped flags, optional WARNINGS_AS_ERRORS)
- fix warning diagnostics across sources:
  - remove unused variables in kmipcore tests and kmippp
  - fix conversion/sign-conversion and cast-qual warnings
  - fix shadowed names and extra semicolons
  - fix struct/class forward-declaration mismatch
  - initialize missing designated fields in tests/examples
- improve OpenSSL compatibility handling in NetClientOpenSSL
  (version-aware certificate getter and method selection)
- add threads linkage in kmipclient CMake for platforms requiring explicit pthread
- silence known third-party clang warning in googletest target setup

Result: clean warning output in maintainer-mode builds for local project code.

+++ kmipcore/include/kmipcore/kmip_enums.hpp
@@ -951,7 +951,7 @@ namespace kmipcore {
     KMIP_REVOKE_SUSPENDED = 0x05,
     KMIP_REVOKE_CESSATION_OF_OPERATION = 0x06,
     KMIP_REVOKE_PRIVILEDGE_WITHDRAWN = 0x07,
-    KMIP_REVOKE_EXTENSIONS = static_cast<std::uint32_t>(0x80000000u)
+    KMIP_REVOKE_EXTENSIONS = 0x80000000u
   };
 
   /** @brief KMIP secret payload data type identifiers. */
@@ -959,7 +959,7 @@ namespace kmipcore {
     // KMIP 1.0
     KMIP_SECDATA_PASSWORD = 0x01,
     KMIP_SECDATA_SEED = 0x02,
-    KMIP_SECDATA_EXTENSIONS = static_cast<std::uint32_t>(0x80000000u)
+    KMIP_SECDATA_EXTENSIONS = 0x80000000u
   };
 
   // ---------------------------------------------------------------------------
